+++ model_builder/edition/views_edition.py
@@ -1,6 +1,7 @@
 import json
 
 from django.http import HttpResponse
+from django.template.loader import render_to_string
 from efootprint.core.hardware.server_base import ServerBase
 
 from model_builder.class_structure import generate_object_edition_structure
@@ -41,16 +42,24 @@ def open_edit_object_panel(request, object_id):
 
 
 @render_exception_modal_if_error
-def edit_object(request, object_id, model_web=None):
-    if model_web is None:
-        model_web = ModelWeb(request.session)
+def edit_object(request, object_id):
+    recompute_modeling = request.POST.get("recomputation", False)
+    model_web = ModelWeb(request.session, launch_system_computations=recompute_modeling,
+                         set_trigger_modeling_updates_to_false=not recompute_modeling)
     obj_to_edit = model_web.get_web_object_from_efootprint_id(object_id)
     if issubclass(obj_to_edit.efootprint_class, ServerBase):
-        storage_data = json.loads(request.POST.get('storage_form_data'))
+        storage_data = json.loads(request.POST.get("storage_form_data"))
         storage = model_web.get_web_object_from_efootprint_id(storage_data["storage_id"])
         edit_object_in_system(storage_data, storage)
+
     response_html, ids_of_web_elements_with_lines_to_remove, data_attribute_updates, top_parent_ids = (
         compute_edit_object_html_and_event_response(request.POST, obj_to_edit))
+
+    if recompute_modeling:
+        refresh_content_response = render_to_string(
+            "model_builder/result/result_panel.html", context={"model_web": model_web})
+        response_html += (f"<div id='result-block' hx-swap-oob='innerHTML:#result-block'>"
+                         f"{refresh_content_response}</div>")
 
     toast_and_highlight_data = {
         "ids": [mirrored_card.web_id for mirrored_card in obj_to_edit.mirrored_cards],
@@ -60,7 +69,7 @@ def edit_object(request, object_id, model_web=None):
 
     return generate_http_response_from_edit_html_and_events(
         response_html, ids_of_web_elements_with_lines_to_remove, data_attribute_updates, top_parent_ids,
-        toast_and_highlight_data)
+       toast_and_highlight_data)
 
 
 def save_model_name(request):