+++ model_builder/modeling_objects_web.py
@@ -90,6 +90,9 @@ class ModelingObjectWeb:
 
         self._modeling_obj.__setattr__(key, value, check_input_validity)
 
+        if self.model_web.launch_system_computations:
+            self.model_web.raise_incomplete_modeling_errors()
+
     def get_efootprint_value(self, key):
         return getattr(self._modeling_obj, key, None)